hitl_oai/pyproject.toml
[project]
name = "oai-agents-arcade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agents-arcade>=0.0.3",
    "python-dotenv>=1.1.0",
]
